+++ pureport/resource_pureport_azure_connection.go
@@ -7,6 +7,7 @@ import (
 	"path/filepath"
 
 	"github.com/antihax/optional"
+	"github.com/hashicorp/terraform-plugin-sdk/helper/customdiff"
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/helper/structure"
 	"github.com/hashicorp/terraform-plugin-sdk/helper/validation"
@@ -66,6 +67,14 @@ func resourceAzureConnection() *schema.Resource {
 			Create: schema.DefaultTimeout(connection.CreateTimeout),
 			Delete: schema.DefaultTimeout(connection.DeleteTimeout),
 		},
+		CustomizeDiff: customdiff.If(
+			customdiff.ResourceConditionFunc(func(d *schema.ResourceDiff, meta interface{}) bool {
+				return d.HasChange("speed") || d.HasChange("high_availability")
+			}),
+			schema.CustomizeDiffFunc(func(d *schema.ResourceDiff, meta interface{}) error {
+				return connection.CloudResourceDiff(d)
+			}),
+		),
 	}
 }
 
@@ -79,13 +88,13 @@ func expandAzureConnection(d *schema.ResourceData) client.AzureExpressRouteConne
 
 	// Create the body of the request
 	c := client.AzureExpressRouteConnection{
-		Type_: "AZURE_EXPRESS_ROUTE",
+		Type:  "AZURE_EXPRESS_ROUTE",
 		Name:  d.Get("name").(string),
 		Speed: int32(speed),
-		Location: &client.Link{
+		Location: client.Link{
 			Href: d.Get("location_href").(string),
 		},
-		Network: &client.Link{
+		Network: client.Link{
 			Href: d.Get("network_href").(string),
 		},
 		BillingTerm: d.Get("billing_term").(string),
@@ -123,7 +132,7 @@ func resourceAzureConnectionCreate(d *schema.ResourceData, m interface{}) error 
 	ctx := config.Session.GetSessionContext()
 
 	opts := client.AddConnectionOpts{
-		Body: optional.NewInterface(c),
+		Connection: optional.NewInterface(c),
 	}
 
 	_, resp, err := config.Session.Client.ConnectionsApi.AddConnection(
@@ -135,7 +144,7 @@ func resourceAzureConnectionCreate(d *schema.ResourceData, m interface{}) error 
 	if err != nil {
 
 		http_err := err
-		json_response := string(err.(client.GenericSwaggerError).Body()[:])
+		json_response := string(err.(client.GenericOpenAPIError).Body()[:])
 		response, err := structure.ExpandJsonFromString(json_response)
 		if err != nil {
 			log.Printf("Error Creating new %s: %v", connection.AzureConnectionName, err)
@@ -203,7 +212,7 @@ func resourceAzureConnectionRead(d *schema.ResourceData, m interface{}) error {
 	d.Set("high_availability", conn.HighAvailability)
 	d.Set("href", conn.Href)
 	d.Set("name", conn.Name)
-	d.Set("peering_type", conn.Peering.Type_)
+	d.Set("peering_type", conn.Peering.Type)
 	d.Set("service_key", conn.ServiceKey)
 	d.Set("speed", conn.Speed)
 	d.Set("state", conn.State)
@@ -247,24 +256,19 @@ func resourceAzureConnectionUpdate(d *schema.ResourceData, m interface{}) error 
 
 	c := expandAzureConnection(d)
 
-	d.Partial(true)
-
 	config := m.(*configuration.Config)
 	ctx := config.Session.GetSessionContext()
 
 	if d.HasChange("name") {
 		c.Name = d.Get("name").(string)
-		d.SetPartial("name")
 	}
 
 	if d.HasChange("description") {
 		c.Description = d.Get("description").(string)
-		d.SetPartial("description")
 	}
 
 	if d.HasChange("speed") {
 		c.Speed = int32(d.Get("speed").(int))
-		d.SetPartial("speed")
 	}
 
 	if d.HasChange("customer_networks") {
@@ -285,7 +289,7 @@ func resourceAzureConnectionUpdate(d *schema.ResourceData, m interface{}) error 
 	}
 
 	opts := client.UpdateConnectionOpts{
-		Body: optional.NewInterface(c),
+		Connection: optional.NewInterface(c),
 	}
 
 	_, resp, err := config.Session.Client.ConnectionsApi.UpdateConnection(
@@ -296,7 +300,7 @@ func resourceAzureConnectionUpdate(d *schema.ResourceData, m interface{}) error 
 
 	if err != nil {
 
-		if swerr, ok := err.(client.GenericSwaggerError); ok {
+		if swerr, ok := err.(client.GenericOpenAPIError); ok {
 
 			json_response := string(swerr.Body()[:])
 			response, jerr := structure.ExpandJsonFromString(json_response)
@@ -319,8 +323,6 @@ func resourceAzureConnectionUpdate(d *schema.ResourceData, m interface{}) error 
 	if err := connection.WaitForConnection(connection.AzureConnectionName, d, m); err != nil {
 		return fmt.Errorf("Error waiting for %s: err=%s", connection.AzureConnectionName, err)
 	}
-
-	d.Partial(false)
 
 	return resourceAzureConnectionRead(d, m)
 }